requirements.txt
google-api-python-client
python-dotenv
simpleaudio
tqdm